requirements_backend.txt
# FastAPI backend
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Database
pymongo>=4.6.0

# Data validation
pydantic>=2.6.0

# File handling (for file scrubber)
python-multipart>=0.0.9

# Testing
pytest>=8.0.0
httpx>=0.27.0
openpyxl>=3.0.10

# Scrubbing
presidio-analyzer
presidio-anonymizer

# OCR
pytesseract
opencv-python-headless

# JWT
pyjwt